Renamed useDeepCompare to useDeepEffect
